+++ src/components/Dashboard.vue
@@ -5,9 +5,9 @@
     </div>
     <v-row class="my-2" v-if="isClient">
       <v-col :sm="12">
-      <div>
-        Here's your activities for this week
-      </div>
+        <div>
+          Here's your activities for this week
+        </div>
         <v-sheet height="500">
           <v-calendar
             ref="calendar"
@@ -19,45 +19,35 @@
           ></v-calendar>
         </v-sheet>
       </v-col>
-      <v-col :sm="6">
-        <div>
-          Steps taken
-        </div>
-        <v-sheet color="success">
-          <v-sparkline
-            :value="steps"
-            color="rgba(255, 255, 255, .7)"
-            fill
-            :height="100"
-            :padding="16"
-            stroke-linecap="round"
-            smooth
-          >
-            <template v-slot:label="item">
-              {{ timeline[item.index] }}
-            </template>
-          </v-sparkline>
-        </v-sheet>
+      <v-col v-if="tagListings1.length" cols="12">
+        <h3>Recommendations for "{{ recommendationTag1.name }}"</h3>
+        <v-slide-group>
+          <listing-card :listing="listing" :small="true" v-for="listing, i in tagListings1" :key="i"></listing-card>
+        </v-slide-group>
       </v-col>
-      <v-col :sm="6">
-        <div>
-          Estd. calories burned
-        </div>
-        <v-sheet color="warning">
-          <v-sparkline
-            :value="calories"
-            color="rgba(255, 255, 255, .7)"
-            fill
-            :height="100"
-            :padding="16"
-            stroke-linecap="round"
-            smooth
-          >
-            <template v-slot:label="item">
-              {{ timeline[item.index] }}
-            </template>
-          </v-sparkline>
-        </v-sheet>
+      <v-col v-if="tagListings2.length" cols="12">
+        <h3>Recommendations for "{{ recommendationTag2.name }}"</h3>
+        <v-slide-group>
+          <listing-card :listing="listing" :small="true" v-for="listing, i in tagListings2" :key="i"></listing-card>
+        </v-slide-group>
+      </v-col>
+      <v-col v-if="tagListings3.length" cols="12">
+        <h3>Recommendations for "{{ recommendationTag3.name }}"</h3>
+        <v-slide-group>
+          <listing-card :listing="listing" :small="true" v-for="listing, i in tagListings3" :key="i"></listing-card>
+        </v-slide-group>
+      </v-col>
+      <v-col v-if="demographicRecommendations.length" cols="12">
+        <h3>Popular among similar users</h3>
+        <v-slide-group>
+          <listing-card :listing="listing" :small="true" v-for="listing, i in demographicRecommendations" :key="i"></listing-card>
+        </v-slide-group>
+      </v-col>
+      <v-col v-if="subscriptionRecommendations.length" cols="12">
+        <h3>Recommended from your subscriptions</h3>
+        <v-slide-group>
+          <listing-card :listing="listing" :small="true" v-for="listing, i in subscriptionRecommendations" :key="i"></listing-card>
+        </v-slide-group>
       </v-col>
     </v-row>
     <v-row class="ma-2" v-else>
@@ -69,8 +59,11 @@
 import moment from 'moment'
 import { getEnrollments } from '../api/plans'
 import { mapGetters } from 'vuex'
+import { getRecommendationsByDemographic, getRecommendationsByTagId, getSubscriptionRecommendations } from '../api/recommendations'
+import ListingCard from './ListingCard.vue'
 
 export default {
+  components: { ListingCard },
   data () {
     return {
       type: 'week',
@@ -82,7 +75,15 @@ export default {
         '#33a02c',
         'primary'
       ],
-      events: []
+      events: [],
+      recommendationTag1: null,
+      recommendationTag2: null,
+      recommendationTag3: null,
+      tagListings1: [],
+      tagListings2: [],
+      tagListings3: [],
+      demographicRecommendations: [],
+      subscriptionRecommendations: []
     }
   },
   computed: {
@@ -96,17 +97,6 @@ export default {
       } else {
         return 'Good Evening'
       }
-    },
-    timeline () {
-      return Array(7).fill()
-        .map((_, i) => moment().subtract(7 - i, 'days').format('M/DD'))
-    },
-    steps () {
-      return Array(7).fill()
-        .map((_, i) => this.rng(1500, 6000))
-    },
-    calories () {
-      return this.steps.map(step => Math.round((step / 1000) * 35) + this.rng(200, 500))
     }
   },
   methods: {
@@ -117,6 +107,8 @@ export default {
   async mounted () {
     this.events.splice(0, this.events.length)
     const enrollments = (await getEnrollments()).data
+    const tagMap = new Map()
+
     for (const sub of enrollments) {
       const subColor = this.colors.pop()
       const startDate = moment(sub.startDate).valueOf()
@@ -129,7 +121,32 @@ export default {
         timed: true
       }))
       this.events.push(...mappedEvents)
+
+      sub.listing.tags.forEach(tag => {
+        tagMap.set(tag.id, tag)
+      })
     }
+
+    const [tagId1, tagId2, tagId3] = Array.from(tagMap.keys()).sort(() => Math.random() - Math.random())
+      .slice(0, 3)
+    this.recommendationTag1 = tagMap.get(tagId1)
+    this.recommendationTag2 = tagMap.get(tagId2)
+    this.recommendationTag3 = tagMap.get(tagId3)
+    console.log(tagId1, tagId2, tagMap)
+
+    const tagListingCollection = [this.tagListings1, this.tagListings2, this.tagListings3];
+    [this.recommendationTag1, this.recommendationTag2, this.recommendationTag3].forEach(async (tag, index) => {
+      if (tag) {
+        const tagListings = (await getRecommendationsByTagId(tag.id)).data
+        tagListingCollection[index].splice(0, tagListingCollection.length, ...tagListings)
+      }
+    })
+
+    const demoListings = (await getRecommendationsByDemographic()).data
+    this.demographicRecommendations.splice(0, this.demographicRecommendations.length, ...demoListings)
+
+    const subListings = (await getSubscriptionRecommendations()).data
+    this.subscriptionRecommendations.splice(0, this.subscriptionRecommendations.length, ...subListings)
   }
 }
 </script>
